+++ src/myoddweb.commandlineparser/commandlineparser.cs
@@ -28,12 +28,6 @@ namespace myoddweb.commandlineparser
   public class CommandlineParser : ICommandlineParser
   {
     /// <summary>
-    /// The patern we want our command line to lead with
-    /// For example, "-hello world" the leading pattern is '-', (it can be more than one char).
-    /// </summary>
-    private readonly string _leadingPattern;
-
-    /// <summary>
     /// A dictionalry with key->value of all the arguments.
     /// </summary>
     private readonly IDictionary<string, string> _parsedArguments;
@@ -53,7 +47,7 @@ namespace myoddweb.commandlineparser
     public CommandlineParser(IReadOnlyList<string> args, ICommandlineArgumentRules commandlineRules = null, string leadingPattern = "--")
     {
       // the leading pattern
-      _leadingPattern = leadingPattern;
+      LeadingPattern = leadingPattern ?? throw new ArgumentNullException( nameof(leadingPattern));
 
       // set the arguments data.
       _commandlineRules = commandlineRules ?? new CommandlineArgumentRules();
@@ -80,7 +74,7 @@ namespace myoddweb.commandlineparser
       var values = new List<string>();
       foreach (var argumentAndKey in _parsedArguments)
       {
-        values.Add($"{_leadingPattern}{argumentAndKey.Key}");
+        values.Add($"{LeadingPattern}{argumentAndKey.Key}");
         if (!(argumentAndKey.Value?.Length > 0))
         {
           continue;
@@ -129,17 +123,22 @@ namespace myoddweb.commandlineparser
     private bool IsKey(string str)
     {
       // check for the leading pattern
-      return str.StartsWith(_leadingPattern);
+      return str.StartsWith(LeadingPattern);
     }
 
-    private static string AdjustedKey(string key)
+    /// <summary>
+    /// Make sure that the given key is lower case and trimmed.
+    /// </summary>
+    /// <param name="key"></param>
+    /// <returns></returns>
+    private static string CleanKeyValue(string key)
     {
       if (null == key)
       {
         // the key cannot be null.
         throw new ArgumentNullException(nameof(key));
       }
-      return key.ToLower();
+      return key.ToLower().Trim();
     }
 
     /// <summary>
@@ -166,7 +165,7 @@ namespace myoddweb.commandlineparser
 
         if (IsKey(args[i]))
         {
-          key = args[i].Substring(_leadingPattern.Length);
+          key = args[i].Substring(LeadingPattern.Length);
 
           if (i + 1 < args.Count && !IsKey(args[i + 1]))
           {
@@ -187,7 +186,7 @@ namespace myoddweb.commandlineparser
         }
         if (key != null)
         {
-          _parsedArguments[AdjustedKey(key)] = val;
+          _parsedArguments[CleanKeyValue(key)] = val;
         }
       }
     }
@@ -195,12 +194,15 @@ namespace myoddweb.commandlineparser
 
     #region  ICommandlineParser
     /// <inheritdoc />
+    public string LeadingPattern { get; }
+
+    /// <inheritdoc />
     public IReadOnlyCommandlineArgumentRules Rules => _commandlineRules;
 
     /// <inheritdoc />
     public ICommandlineParser Clone()
     {
-      return new CommandlineParser(Arguments(false), _commandlineRules, _leadingPattern);
+      return new CommandlineParser(Arguments(false), _commandlineRules, LeadingPattern);
     }
 
     /// <inheritdoc />
@@ -212,7 +214,7 @@ namespace myoddweb.commandlineparser
       }
 
       // adjust the key value
-      var adjustedKey = AdjustedKey(key);
+      var adjustedKey = GetAdjustedKeyFromActualOrAliasKey(key);
 
       // remove the key
       _parsedArguments.Remove(adjustedKey);
@@ -233,8 +235,7 @@ namespace myoddweb.commandlineparser
     /// <inheritdoc />
     public T Get<T>(string key)
     {
-      var adjustedKey = AdjustedKey(key);
-      var s = Get(adjustedKey);
+      var s = Get(key);
       if (s == null)
       {
         return default(T);
@@ -253,8 +254,7 @@ namespace myoddweb.commandlineparser
     /// <inheritdoc />
     public T Get<T>(string key, T defaultValue)
     {
-      var adjustedKey = AdjustedKey(key);
-      var s = Get(adjustedKey);
+      var s = Get(key);
       if (s == null)
       {
         return defaultValue;
@@ -274,7 +274,7 @@ namespace myoddweb.commandlineparser
     public string Get(string key, string defaultValue)
     {
       // adjust the key value
-      var adjustedKey = AdjustedKey(key);
+      var adjustedKey = GetAdjustedKeyFromActualOrAliasKey(key);
 
       // return it if we have it, the default otherwise.
       return !_parsedArguments.ContainsKey(adjustedKey) ? defaultValue : _parsedArguments[adjustedKey];
@@ -284,24 +284,60 @@ namespace myoddweb.commandlineparser
     public string Get(string key)
     {
       // adjust the key value
-      var adjustedKey = AdjustedKey(key);
+      var adjustedKey = GetAdjustedKeyFromActualOrAliasKey(key);
       if (_parsedArguments.ContainsKey(adjustedKey))
       {
         return _parsedArguments[adjustedKey];
       }
 
       // look for that key in our default arguments.
-      var defaultValue = _commandlineRules.FirstOrDefault(r => r.Keys.Contains(adjustedKey));
+      var possible = _commandlineRules.FirstOrDefault(r => r.Keys.Contains(adjustedKey));
 
       // return it if we have it, the default otherwise.
-      return defaultValue?.DefaultValue;
+      return possible?.DefaultValue;
+    }
+
+    /// <summary>
+    /// Given a possible key value we get the actual value of they key
+    /// That matches what the user gave us.
+    /// </summary>
+    /// <param name="actualOrAliasKey"></param>
+    /// <returns></returns>
+    private string GetAdjustedKeyFromActualOrAliasKey( string actualOrAliasKey )
+    {
+      var adjustedActualOrAliasKey = CleanKeyValue(actualOrAliasKey);
+
+      // look for that key in our default arguments.
+      var possible = _commandlineRules.FirstOrDefault(r => r.Keys.Contains(adjustedActualOrAliasKey));
+
+      // if we cannot find that key
+      // then we assume that the 'actual' key is what was given to us
+      // and we just return the adjusted value
+      if (null == possible)
+      {
+        return adjustedActualOrAliasKey;
+      }
+
+      // find the first adjusted value that exists in the arguments
+      foreach (var possibleKey in possible.Keys)
+      {
+        var possibleAdjustedKey = CleanKeyValue(possibleKey);
+        if (_parsedArguments.ContainsKey(possibleAdjustedKey))
+        {
+          return possibleAdjustedKey;
+        }
+      }
+
+      // we could not find that key at all
+      // so we do not have any adjusted and/or actual value.
+      return adjustedActualOrAliasKey;
     }
 
     /// <inheritdoc />
     public bool IsSet(string key)
     {
       // adjust the key value
-      var adjustedKey = AdjustedKey(key);
+      var adjustedKey = GetAdjustedKeyFromActualOrAliasKey(key);
 
       // look if the value is simply set this is the default.
       if (_parsedArguments.ContainsKey(adjustedKey))